Full project reorg.
