+++ mac/solvers/greedy_eig.py
