+++ mac/utils/graphs.py
@@ -1,139 +1,14 @@
+"""
+Types and utilities for working with graphs.
+"""
+
 import numpy as np
-from collections import namedtuple
-import matplotlib.pyplot as plt
-from scipy.sparse import csr_matrix, coo_matrix, csc_matrix
-import networkx as nx
-import math
 from typing import List, Union, Tuple
+from collections import namedtuple
+from scipy.sparse import csr_matrix, coo_matrix, csc_matrix
 
 # Define Edge container
 Edge = namedtuple("Edge", ["i", "j", "weight"])
-
-def round_nearest(w, k, weights=None, break_ties_decimal_tol=None):
-    """
-    Round a solution w to the relaxed problem, i.e. w \in [0,1]^m, |w| = k to a
-    solution to the original problem with w_i \in {0,1}. Ties between edges
-    are broken based on the original weight (all else being equal, we
-    prefer edges with larger weight).
-
-    w: A solution in the feasible set for the relaxed problem
-    weights: The original weights of the edges to use as a tiebreaker
-    k: The number of edges to select
-    break_ties_decimal_tol: tolerance for determining floating point equality of weights w. If two selection weights are equal to this many decimal places, we break the tie based on the original weight.
-
-    returns w': A solution in the feasible set for the original problem
-    """
-    if weights is None or break_ties_decimal_tol is None:
-        # If there are no tiebreakers, just set the top k elements of w to 1,
-        # and the rest to 0
-        idx = np.argpartition(w, -k)[-k:]
-        rounded = np.zeros(len(w))
-        if k > 0:
-            rounded[idx] = 1.0
-        return rounded
-
-    # If there are tiebreakers, we truncate the selection weights to the
-    # specified number of decimal places, and then break ties based on the
-    # original weights
-    truncated_w = w.round(decimals=break_ties_decimal_tol)
-    zipped_vals = np.array(
-        [(truncated_w[i], weights[i]) for i in range(len(w))],
-        dtype=[("w", "float"), ("weight", "float")],
-    )
-    idx = np.argpartition(zipped_vals, -k, order=["w", "weight"])[-k:]
-    rounded = np.zeros(len(w))
-    if k > 0:
-        rounded[idx] = 1.0
-    return rounded
-
-def round_random(w, k):
-    """
-    Round a solution w to the relaxed problem, i.e. w \in [0,1]^m, |w| = k to
-    one with hard edge constraints and satisfying the constraint that the
-    expected number of selected edges is equal to k.
-
-    w: A solution in the feasible set for the relaxed problem
-    k: The number of edges to select _in expectation_
-
-    returns w': A solution containing hard edge selections with an expected
-    number of selected edges equal to k.
-    """
-    x = np.zeros(len(w))
-    for i in range(len(w)):
-        r = np.random.rand()
-        if w[i] > r:
-            x[i] = 1.0
-    return x
-
-def round_madow(w, k, seed=None, value_fn=None, max_iters=1):
-    if value_fn is None or max_iters == 1:
-        return round_madow_base(w, k, seed)
-
-    best_x = None
-    best_val = -np.inf
-    for i in range(max_iters):
-        x = round_madow_base(w, k, seed)
-        val = value_fn(x)
-        if val > best_val:
-            best_val = val
-            best_x = x
-    return best_x
-
-
-def round_madow_base(w, k, seed=None):
-    """
-    Use Madow rounding
-    """
-    if seed is None:
-        u = np.random.rand()
-    else:
-        u = seed.rand()
-    x = np.zeros(len(w))
-    # pi = np.zeros(len(w) + 1)
-    pi = np.zeros(len(w))
-    sumw = np.cumsum(w)
-    pi[1:] = sumw[:-1]
-    for i in range(k):
-        total = u + i
-        x[np.where((pi <= total) & (total < sumw))] = 1.0
-        # for j in range(len(w)):
-        #     if (x[j] != 1) and (pi[j] <= total) and (total < pi[j+1]):
-        #         x[j] = 1.0
-        #         break
-
-    assert np.sum(x) == k, f"Error: {np.sum(x)} != {k}"
-    return x
-
-def nx_to_mac(G: nx.Graph) -> List[Edge]:
-    """Returns the list of edges in the graph G
-
-    Args:
-        G (nx.Graph): the graph
-
-    Returns:
-        List[Edge]: the list of edges
-    """
-    edges = []
-    for nxedge in G.edges():
-        edge = Edge(nxedge[0], nxedge[1], 1.0)
-        edges.append(edge)
-    return edges
-
-
-def mac_to_nx(edges: List[Edge]) -> nx.Graph:
-    """returns the graph corresponding to the list of edges
-
-    Args:
-        edges (List[Edge]): the list of edges
-
-    Returns:
-        nx.Graph: the networkx graph
-    """
-    G = nx.Graph()
-    for edge in edges:
-        G.add_edge(edge.i, edge.j, weight=edge.weight)
-    return G
-
 
 def weight_graph_lap_from_edge_list(edges: List[Edge], num_nodes: int) -> csr_matrix:
     """Returns the (sparse) weighted graph Laplacian matrix from the list of edges
@@ -332,4 +207,3 @@ def get_edge_selection_as_binary_mask(
         if edge in selected_edges:
             mask[i] = 1.0
     return mask
-
